requirements.txt
cassandra-driver==3.12.1
robotframework==3.0.2